% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawDensityPlot.R
\name{drawDensityPlot}
\alias{drawDensityPlot}
\title{Draw Density Plot}
\usage{
drawDensityPlot(
  metadata,
  cellType,
  widthValue = 4,
  heightValue = 9.8,
  resultsFolder = resultsDirectory
)
}
\arguments{
\item{metadata}{seurat data object}

\item{cellType}{name of dataset to use on the saved objects}

\item{widthValue}{width of the final plot (default 6 cm)}

\item{heightValue}{height of the final plot (default 6 cm)}
}
\description{
Function that generates density plots for a given data
}
