% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawDimPlot.R
\name{drawDimPlot}
\alias{drawDimPlot}
\title{Draw Dim Plot}
\usage{
drawDimPlot(
  seuratObject,
  datasetName,
  mapType = "umap",
  splitValue = NULL,
  heightValue = 10,
  widthValue = 10,
  colorVector = NULL,
  cells2highlight = NULL,
  sizeHighlightedCells = NULL,
  groupbyValue = NULL,
  labelValue = F,
  showLegend = FALSE,
  resultsFolder = resultsDirectory,
  plotMinimal = FALSE
)
}
\arguments{
\item{seuratObject}{seurat data object}

\item{datasetName}{name of dataset to use on the saved objects}

\item{mapType}{set umap as default}

\item{splitValue}{if the umap is to be splitted}

\item{heightValue}{height of the final plot (default 6 cm)}

\item{widthValue}{width of the final plot (default 6 cm)}

\item{colorVector}{}

\item{cells2highlight}{highlight specific cells}

\item{sizeHighlightedCells}{change size of highlighted cells}

\item{groupbyValue}{metadava variable to visualize in the dim plot}

\item{labelValue}{show or hide cluster labels (default FALSE)}

\item{showLegend}{show or hide plot legend (default FALSE)}

\item{resultsFolder}{directory where to save the results}

\item{plotMinimal}{show a minimal plot}
}
\description{
Function that generates dim plots for a given Seurat dataset
}
