% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawDotPlot.R
\name{drawDotPlot}
\alias{drawDotPlot}
\title{Draw Dot Plot}
\usage{
drawDotPlot(
  dataset,
  genesInterest,
  splitby = NULL,
  plotName,
  widthValue = 10,
  heightValue = 10
)
}
\arguments{
\item{dataset}{seurat data object}

\item{genesInterest}{define which features (genes) to show}

\item{splitby}{}

\item{plotName}{define name of the plot}

\item{widthValue}{width of the final plot (default 6 cm)}

\item{heightValue}{height of the final plot (default 6 cm)}
}
\description{
Function that generates dot plots for a given Seurat dataset
}
