% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawFeaturePlot.R
\name{drawFeaturePlot}
\alias{drawFeaturePlot}
\title{Draw Feature Plot}
\usage{
drawFeaturePlot(
  seuratObject,
  featureNames,
  datasetName,
  reductionValue = "umap",
  widthValue = 5,
  heightValue = 5,
  pointSizeValue = 0.3,
  nrowValue = 1,
  ncolValue = 1
)
}
\arguments{
\item{seuratObject}{seurat data object}

\item{featureNames}{name of the features}

\item{datasetName}{name of dataset to use on the saved objects}

\item{reductionValue}{which dimensionality reduction method to use (default umap)}

\item{pointSizeValue}{size of the points (default 0.3)}

\item{nrowValue}{number of rows (default 1 for only one feature)}

\item{ncolValue}{number of colums (default 1 for only one feature)}
}
\description{
Function that generates feature plots for given features
}
