% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawHeatmapPlot.R
\name{drawHeatmapPlot}
\alias{drawHeatmapPlot}
\title{Draw Heatmap Plot}
\usage{
drawHeatmapPlot(
  seuratObject,
  featureNames,
  widthValue = 28,
  heightValue = 15,
  plotName,
  assaytype = "SCT",
  groupValue = "orig.ident",
  slotValue = "scale.data",
  drawLinesValue = T,
  cellsValue = NULL
)
}
\arguments{
\item{seuratObject}{seurat data object}

\item{featureNames}{list of features to plot}

\item{widthValue}{width of the final plot (default 28 cm)}

\item{heightValue}{height of the final plot (default 15 cm)}

\item{plotName}{name of the plot}

\item{assaytype}{TEST}

\item{groupValue}{TEST}

\item{slotValue}{TEST}

\item{drawLinesValue}{TEST}

\item{cellsValue}{TEST}

\item{datasetName}{name of the dataset (e.g., Pestana2022)}
}
\description{
Function that generates heamap plots for given features
}
