% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawSpatialDimPlot.R
\name{drawSpatialDimPlot}
\alias{drawSpatialDimPlot}
\title{Draw Spatial Plot}
\usage{
drawSpatialDimPlot(
  spatialObject,
  groupValue,
  recursiveSection = TRUE,
  sectionID = NULL
)
}
\arguments{
\item{spatialObject}{spatial Seurat object to use}

\item{groupValue}{metadata variable to show on the plot}

\item{recursiveSection}{if plots should be made for all sections (default TRUE)}

\item{sectionID}{if recursiveSection is FALSE, then a specific section ID should be defined (default NULL)}
}
\value{
This function generates and saves a spatial Dim plot in the Results folder and prints in the screen.
}
\description{
Function that generates spatial dim plots for a given Seurat dataset
}
\examples{
drawSpatialDimPlot(spatialObject = spatial,groupValue = "normCTXDistance_Bin5",recursiveSection = TRUE )
drawSpatialDimPlot(spatialObject = spatial,groupValue = "normCTXDistance_Bin5",recursiveSection = FALSE,sectionID = "B1_A1" )

}
