% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawSpatialFeaturePlot.R
\name{drawSpatialFeaturePlot}
\alias{drawSpatialFeaturePlot}
\title{Draw Spatial Feature Plot for given genes}
\usage{
drawSpatialFeaturePlot(features, dataset, metadata, subsample = FALSE)
}
\arguments{
\item{dataset}{TEST}
}
\description{
Function that generates spatial feature plots for a given gene and sample
}
