% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawSpatialDim.R
\name{drawSpatialPlot}
\alias{drawSpatialPlot}
\title{Draw Spatial Plot}
\usage{
drawSpatialPlot(features, dataset, metadata, subsample = FALSE)
}
\arguments{
\item{dataset}{TEST}
}
\description{
Function that generates dot plots for a given Seurat dataset
}
