% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawVlnPlot.R
\name{drawVlnPlot}
\alias{drawVlnPlot}
\title{Draw Violin Plot}
\usage{
drawVlnPlot(
  dataset,
  feature = featureList,
  plotName = featureGroup,
  orderIdentity = orderIdentity,
  identity = identityValue,
  splitby = NULL,
  datasetName,
  pointSize = 0.1,
  colorVector = NULL,
  heightValue = 6,
  widthValue = 12,
  nrowValue = 2,
  ncolValue = 1,
  fontsizeValueX = 15,
  fontsizeValueY = 15,
  addXlabel = T,
  resultsFolder = resultsDirectory
)
}
\arguments{
\item{dataset}{seurat data object}

\item{feature}{list of features to be plotted}

\item{splitby}{if violins are to be splitted between an indicated metadata variable}

\item{pointSize}{define the size of the dots in the plot. 0 indicates no points should be shown}

\item{colorVector}{vector of colors to manual color each violin}

\item{heightValue}{height of the final plot (default 6 cm)}

\item{widthValue}{width of the final plot (default 6 cm)}

\item{nrowValue}{number of rows in the final panel}

\item{ncolValue}{number of columns in the final panel}

\item{addXlabel}{if x axis ticks should be labeled}

\item{fontsizeValue}{size of text in y axis}
}
\description{
Function that generates violin plots for N number of indicated features
}
