% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savePlots.R
\name{savePlots}
\alias{savePlots}
\title{Function that saves plots}
\usage{
savePlots(
  plotObject,
  plotTypeValue,
  fileFormat = c(".png", ".pdf"),
  widthValue,
  heightValue,
  measureUnitsValue = "cm",
  extraInfoValue
)
}
\arguments{
\item{plotObject}{plot object generated by each draw plot function}

\item{plotTypeValue}{name of type of plot used (name defined by default in each draw plot function)}

\item{fileFormat}{format of the files to be generated (default .png and .pdf)}

\item{widthValue}{width of the plot (defined by default in each draw plot function)}

\item{heightValue}{height of the plot (defined by default in each draw plot function)}

\item{measureUnitsValue}{measure units of the plot (default cm)}

\item{extraInfoValue}{extra information to be appended to the file name}
}
\description{
Function that saves plots in multiple formats
}
