% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawDotPlot.R
\name{drawDotPlot}
\alias{drawDotPlot}
\title{Draw Dot Plot}
\usage{
drawDotPlot(
  seuratObject,
  plotName,
  featureValues,
  splitby = NULL,
  widthValue = 10,
  heightValue = 10
)
}
\arguments{
\item{seuratObject}{seurat data object}

\item{plotName}{define name of the plot}

\item{featureValues}{define which features (genes) to show}

\item{splitby}{metadata variable to split (default NULL)}

\item{widthValue}{width of the final plot (default 6 cm)}

\item{heightValue}{height of the final plot (default 6 cm)}
}
\description{
Function that generates dot plots for a given Seurat dataset
}
\examples{
drawDotPlot(seuratObject = seuratLOG,plotName = "test",featureValues = c("S100A9","NKG7","LDHB","CD79A"))


}
