% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawFeaturePlot.R
\name{drawFeaturePlot}
\alias{drawFeaturePlot}
\title{Draw Feature Plot}
\usage{
drawFeaturePlot(
  seuratObject,
  featureValues,
  datasetName,
  reductionValue = "umap",
  widthValue = 7,
  heightValue = 7,
  pointSizeValue = 0.3,
  nrowValue = 1,
  ncolValue = 1
)
}
\arguments{
\item{seuratObject}{seurat data object}

\item{featureValues}{name of the features}

\item{datasetName}{name of dataset to use on the saved objects}

\item{reductionValue}{which dimensionality reduction method to use (default umap)}

\item{widthValue}{width of the final plot (default 7 cm)}

\item{heightValue}{height of the final plot (default 7 cm)}

\item{pointSizeValue}{size of the points (default 0.3)}

\item{nrowValue}{number of rows (default 1 for only one feature)}

\item{ncolValue}{number of colums (default 1 for only one feature)}
}
\description{
Function that generates feature plots for given features
}
\examples{
drawFeaturePlot(seuratObject = seuratLOG,feature = c("S100A9","NKG7","LDHB","CD79A"),nrowValue = 2,ncolValue = 2,datasetName = "test")
}
